# api/openapi/openapi-3.0.yaml
openapi: 3.0.3
info:
  title: LinguaFlow Web Service API
  version: 0.1.0
servers:
  - url: /api/v1
paths:
...
  /projects:
    post:
      operationId: CreateProject
      summary: 创建项目
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        default:
          $ref: '#/components/responses/Problem'
    get:
      operationId: ListProjects
      summary: 列出项目
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 项目列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectListResponse'
        default:
          $ref: '#/components/responses/Problem'
  /projects/{projectId}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      operationId: GetProject
      summary: 获取项目详情
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 项目详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        default:
          $ref: '#/components/responses/Problem'
    put:
      operationId: UpdateProject
      summary: 更新项目
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      operationId: DeleteProject
      summary: 删除项目
      security:
        - bearerAuth: []
      responses:
        '204':
          description: 删除成功
        default:
          $ref: '#/components/responses/Problem'
  /orgs/{orgId}/projects:
    parameters:
      - $ref: '#/components/parameters/OrgId'
    post:
      operationId: CreateOrgProject
      summary: 创建组织项目
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        default:
          $ref: '#/components/responses/Problem'
    get:
      operationId: ListOrgProjects
      summary: 列出组织项目
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 项目列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectListResponse'
        default:
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    OrgId:
      name: orgId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
...
  responses:
    Problem:
      description: 错误响应
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      required:
        - title
        - status
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
...
    Project:
      type: object
      required:
        - id
        - name
        - source_lang
        - target_lang
        - glossary_enabled
      properties:
        id:
          type: integer
        name:
          type: string
        owner_user_id:
          type: integer
        owner_org_id:
          type: integer
        config:
          type: object
          additionalProperties: true
        glossary_enabled:
          type: boolean
          default: false
          description: 是否启用术语表
        source_lang:
          type: string
        target_lang:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProjectListResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Project'
    CreateProjectRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        config:
          type: object
          additionalProperties: true
        glossary_enabled:
          type: boolean
          default: false
          description: 是否启用术语表
        source_lang:
          type: string
        target_lang:
          type: string
      additionalProperties: false
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
        config:
          type: object
          additionalProperties: true
        glossary_enabled:
          type: boolean
          description: 是否启用术语表
        source_lang:
          type: string
        target_lang:
          type: string
      additionalProperties: false
...